+++ src/Exception/AutoloadException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Mtarld\SymbokBundle\Exception;
+
+class AutoloadException extends \RuntimeException
+{
+}