+++ tests/Autoload/AutoloaderTest.php
@@ -2,12 +2,13 @@
 
 namespace Mtarld\SymbokBundle\Tests\Autoload;
 
-use Mtarld\SymbokBundle\Autoload\Autoload;
+use App\Entity\Product1;
+use App\Entity\Product2;
+use App\Model\ProductFromAnotherNamespace;
+use Mtarld\SymbokBundle\Autoload\Autoloader;
 use Mtarld\SymbokBundle\Autoload\AutoloadFinder;
+use Mtarld\SymbokBundle\Cache\RuntimeClassCache;
 use Mtarld\SymbokBundle\Replacer\ReplacerInterface;
-use Mtarld\SymbokBundle\Tests\Fixtures\App\src\Entity\Product1;
-use Mtarld\SymbokBundle\Tests\Fixtures\App\src\Entity\Product2;
-use Mtarld\SymbokBundle\Tests\Fixtures\App\src\Model\ProductFromAnotherNamespace;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Filesystem\Filesystem;
@@ -16,13 +17,13 @@ use Symfony\Component\Filesystem\Filesystem;
  * @group unit
  * @group autoload
  */
-class AutoloadTest extends TestCase
+class AutoloaderTest extends TestCase
 {
     /**
      * @dataProvider substituteClassDataProvider
      * @testdox Substitute $class and expect substituted to be $substitute
      */
-    public function testSubstituteClass(string $class, bool $substitute): void
+    public function testSubstituteClass(string $classFqcn, bool $substitute): void
     {
         $logger = $this->createMock(LoggerInterface::class);
         $replacer = $this->createMock(ReplacerInterface::class);
@@ -32,20 +33,19 @@ class AutoloadTest extends TestCase
         ;
 
         $fileSystem = new Filesystem();
-        $oldCacheFilePath = sprintf('var/cache/%s.php', $class);
+        $oldCacheFilePath = sprintf('%s/%s.php', 'var/cache/test/symbok', str_replace('\\', '/', $classFqcn));
         $fileSystem->remove($oldCacheFilePath);
 
         $autoloadFinder = new AutoloadFinder('Mtarld\\SymbokBundle\\Tests\\Fixtures\\App\\src\\Entity');
 
-        $autoload = new Autoload(
+        $autoload = new Autoloader(
             $replacer,
             $logger,
             $autoloadFinder,
-            ['Mtarld\\SymbokBundle\\Tests\\Fixtures\\App\\src\\Entity'],
-            'var/cache/',
-            true
+            new RuntimeClassCache('var/cache/test/symbok/', true),
+            ['App\\Entity']
         );
-        $autoload->loadClass($class);
+        $autoload->loadClass($classFqcn);
     }
 
     public function substituteClassDataProvider(): iterable
@@ -54,6 +54,6 @@ class AutoloadTest extends TestCase
         yield [Product2::class, true];
         yield [ProductFromAnotherNamespace::class, false];
         yield ['AnotherClass', false];
-        yield ['Mtarld\SymbokBundle\Tests\Fixtures\App\src\Entity\VirtualClass', false];
+        yield ['App\Entity\VirtualClass', false];
     }
 }