+++ tests/Fixtures/App/src/Model/ProductFromAnotherNamespace.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Mtarld\SymbokBundle\Tests\Fixtures\App\src\Model;
+namespace App\Model;
 
 class ProductFromAnotherNamespace
 {