+++ CHANGELOG.md
@@ -5,7 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.2.0] - Unreleased
+
+### Added
+
+- Added Doctrine bundle dev dependencies
+- Added doctrine metadata original path by symbok one's replacement
+- Added runtime class cache warmer
+
+### Changed
+
+- Merged unit and functional tests in Makefile
+- Improved fixtures namespace
+- Improved fixtures config files
+
+### Removed
 
 ## [2.1.2] - 2020-05-10